# lib/fs: Optimize WindowsInvalidFilename