@imsodin 2022-11-21
> This will have to wait because we're still targeting 1.18 until 1.20 comes out.

Seems to be a go1.18 change - CI for that version came up green.