@calmh 2022-11-21
Oh, indeed. I had a feeling it was Go 1.19 and discounted the check because I had another feeling this was Windows-only code and not built... But I was wrong on all accounts :D